Group object runner/worker/invoker classes into invokers package
