+++ src/main/java/org/testng/SuiteRunner.java
@@ -8,8 +8,12 @@ import org.testng.collections.Lists;
 import org.testng.collections.Maps;
 import org.testng.collections.Sets;
 import org.testng.internal.*;
-import org.testng.internal.ConfigMethodArguments.Builder;
+import org.testng.internal.invokers.ConfigMethodArguments;
+import org.testng.internal.invokers.ConfigMethodArguments.Builder;
 import org.testng.internal.annotations.IAnnotationFinder;
+import org.testng.internal.invokers.IInvocationStatus;
+import org.testng.internal.invokers.IInvoker;
+import org.testng.internal.invokers.InvokedMethod;
 import org.testng.internal.thread.ThreadUtil;
 import org.testng.reporters.JUnitXMLReporter;
 import org.testng.reporters.TestHTMLReporter;
