+++ src/main/java/org/testng/TestRunner.java
@@ -18,11 +18,11 @@ import java.util.concurrent.atomic.AtomicReference;
 import org.testng.collections.Lists;
 import org.testng.collections.Maps;
 import org.testng.collections.Sets;
-import org.testng.internal.AbstractParallelWorker;
+import org.testng.internal.invokers.AbstractParallelWorker;
 import org.testng.internal.Attributes;
 import org.testng.internal.ClassInfoMap;
-import org.testng.internal.ConfigMethodArguments;
-import org.testng.internal.ConfigMethodArguments.Builder;
+import org.testng.internal.invokers.ConfigMethodArguments;
+import org.testng.internal.invokers.ConfigMethodArguments.Builder;
 import org.testng.internal.ConfigurationGroupMethods;
 import org.testng.internal.DefaultListenerFactory;
 import org.testng.internal.DynamicGraphHelper;
@@ -30,10 +30,10 @@ import org.testng.internal.GroupsHelper;
 import org.testng.internal.IConfigEavesdropper;
 import org.testng.internal.IConfiguration;
 import org.testng.internal.IContainer;
-import org.testng.internal.IInvoker;
+import org.testng.internal.invokers.IInvoker;
 import org.testng.internal.ITestClassConfigInfo;
 import org.testng.internal.ITestResultNotifier;
-import org.testng.internal.Invoker;
+import org.testng.internal.invokers.Invoker;
 import org.testng.internal.RuntimeBehavior;
 import org.testng.internal.TestMethodContainer;
 import org.testng.internal.TestMethodComparator;
@@ -43,7 +43,7 @@ import org.testng.internal.ResultMap;
 import org.testng.internal.RunInfo;
 import org.testng.internal.Systematiser;
 import org.testng.internal.TestListenerHelper;
-import org.testng.internal.TestMethodWorker;
+import org.testng.internal.invokers.TestMethodWorker;
 import org.testng.internal.TestNGClassFinder;
 import org.testng.internal.TestNGMethodFinder;
 import org.testng.internal.Utils;
