+++ src/main/java/org/testng/internal/invokers/ExpectedExceptionsHolder.java
@@ -1,4 +1,4 @@
-package org.testng.internal;
+package org.testng.internal.invokers;
 
 import org.testng.IExpectedExceptionsHolder;
 import org.testng.ITestNGMethod;
