+++ src/main/java/org/testng/internal/invokers/GroupConfigMethodArguments.java
@@ -1,7 +1,8 @@
-package org.testng.internal;
+package org.testng.internal.invokers;
 
 import java.util.Map;
 import org.testng.ITestNGMethod;
+import org.testng.internal.ConfigurationGroupMethods;
 import org.testng.xml.XmlSuite;
 
 public class GroupConfigMethodArguments extends Arguments {
