+++ src/main/java/org/testng/internal/invokers/IInvocationStatus.java
@@ -1,4 +1,4 @@
-package org.testng.internal;
+package org.testng.internal.invokers;
 
 /**
  * Helps keep track of when a method was invoked
