+++ src/main/java/org/testng/internal/invokers/InstanceBasedParallelParallelWorker.java
@@ -1,10 +1,12 @@
-package org.testng.internal;
+package org.testng.internal.invokers;
 
 import org.testng.IMethodInstance;
 import org.testng.ITestNGMethod;
 import org.testng.collections.ListMultiMap;
 import org.testng.collections.Lists;
 import org.testng.collections.Maps;
+import org.testng.internal.MethodHelper;
+import org.testng.internal.invokers.*;
 import org.testng.thread.IWorker;
 
 import java.util.List;
