+++ src/main/java/org/testng/internal/thread/graph/SuiteWorkerFactory.java
@@ -3,7 +3,7 @@ package org.testng.internal.thread.graph;
 import org.testng.ISuite;
 import org.testng.SuiteRunnerWorker;
 import org.testng.collections.Lists;
-import org.testng.internal.SuiteRunnerMap;
+import org.testng.internal.invokers.SuiteRunnerMap;
 
 import java.util.List;
 import org.testng.thread.IThreadWorkerFactory;
